Changed project file structure. Updated imports.
