+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".root.RootActivity"
+            android:name=".features.root.RootActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan">
             <intent-filter>
@@ -23,9 +23,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".login.LoginActivity"/>
-        <activity android:name=".homepage.HomeActivity"/>
-        <activity android:name=".signup.SignupActivity"/>
+        <activity android:name=".features.auth.login.LoginActivity"/>
+        <activity android:name=".features.homepage.HomeActivity"/>
+        <activity android:name=".features.auth.signup.SignupActivity"/>
         <activity android:name=".example.ui.example.ExampleActivity"/>
         <activity
             android:name=".example.ui.viewpager.ViewPagerActivity"
