+++ app/src/main/java/ar/com/wolox/android/example/ui/example/ExamplePresenter.kt
@@ -1,6 +1,6 @@
 package ar.com.wolox.android.example.ui.example
 
-import ar.com.wolox.android.example.utils.UserSession
+import ar.com.wolox.android.utils.UserSession
 import ar.com.wolox.wolmo.core.presenter.BasePresenter
 import javax.inject.Inject
