+++ app/src/main/java/ar/com/wolox/android/example/ui/viewpager/fragment/ViewPagerPresenter.kt
@@ -1,6 +1,6 @@
 package ar.com.wolox.android.example.ui.viewpager.fragment
 
-import ar.com.wolox.android.example.utils.UserSession
+import ar.com.wolox.android.utils.UserSession
 import ar.com.wolox.wolmo.core.presenter.BasePresenter
 import javax.inject.Inject
