+++ app/src/main/java/ar/com/wolox/android/features/auth/signup/SignupActivity.kt
@@ -1,4 +1,4 @@
-package ar.com.wolox.android.homepage
+package ar.com.wolox.android.features.auth.signup
 
 import android.content.Context
 import android.content.Intent
@@ -6,17 +6,16 @@ import ar.com.wolox.android.R
 import ar.com.wolox.android.databinding.ActivityBaseBinding
 import ar.com.wolox.wolmo.core.activity.WolmoActivity
 
-class HomeActivity : WolmoActivity<ActivityBaseBinding>() {
+class SignupActivity : WolmoActivity<ActivityBaseBinding>() {
 
-    override fun layout() = R.layout.home_activity
+    override fun layout() = R.layout.sign_up_activity
 
     override fun init() {
     }
 
     companion object {
-
         fun start(context: Context) {
-            with(Intent(context, HomeActivity::class.java)) {
+            with(Intent(context, SignupActivity::class.java)) {
                 context.startActivity(this)
             }
         }
