+++ app/src/main/java/ar/com/wolox/android/features/homepage/HomeActivity.kt
@@ -1,4 +1,4 @@
-package ar.com.wolox.android.signup
+package ar.com.wolox.android.features.homepage
 
 import android.content.Context
 import android.content.Intent
@@ -6,16 +6,17 @@ import ar.com.wolox.android.R
 import ar.com.wolox.android.databinding.ActivityBaseBinding
 import ar.com.wolox.wolmo.core.activity.WolmoActivity
 
-class SignupActivity : WolmoActivity<ActivityBaseBinding>() {
+class HomeActivity : WolmoActivity<ActivityBaseBinding>() {
 
-    override fun layout() = R.layout.sign_up_activity
+    override fun layout() = R.layout.home_activity
 
     override fun init() {
     }
 
     companion object {
+
         fun start(context: Context) {
-            with(Intent(context, SignupActivity::class.java)) {
+            with(Intent(context, HomeActivity::class.java)) {
                 context.startActivity(this)
             }
         }
