+++ app/src/main/java/ar/com/wolox/android/features/root/RootActivity.kt
@@ -1,10 +1,10 @@
-package ar.com.wolox.android.root
+package ar.com.wolox.android.features.root
 
 import ar.com.wolox.android.R
 import ar.com.wolox.android.databinding.ActivityBaseBinding
-import ar.com.wolox.android.example.utils.UserSession
-import ar.com.wolox.android.homepage.HomeActivity
-import ar.com.wolox.android.login.LoginActivity
+import ar.com.wolox.android.utils.UserSession
+import ar.com.wolox.android.features.homepage.HomeActivity
+import ar.com.wolox.android.features.auth.login.LoginActivity
 import ar.com.wolox.wolmo.core.activity.WolmoActivity
 import javax.inject.Inject
